refactor(utils): restructure function utils
